+++ config.example.js
@@ -1,3 +1,6 @@
+/**
+ * X-Super-Properties header for Discord. This **can be** filled with random values (note: *NOT KEYS*); it has no effect on functionality.
+ */
 const super_properties = Buffer.from(JSON.stringify({
     'os': 'Windows',
     'browser': 'Firefox',
@@ -11,14 +14,16 @@ const super_properties = Buffer.from(JSON.stringify({
     'referring_domain_current':'',
     'release_channel':'stable',
     'client_build_number':51863,
-    'client_event_source':null
-})).toString('base64');
+    'client_event_source':null}
+)).toString('base64');
 
-module.exports = {
-    useragent: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:72.0) Gecko/20100101 Firefox/71.0',
+const useragent = 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:72.0) Gecko/20100101 Firefox/72.0';
+const smspva = '';
+const captcha = '';
 
-    super_properties: super_properties,
-    
-    smspva: '',
-    captcha: ''
+export {
+    useragent,
+    super_properties,
+    smspva,
+    captcha
 }