+++ src/email.js
@@ -1,38 +1,8 @@
-const fetch = require('node-fetch');
-const { format } = require('util');
+import fetch from 'node-fetch';
 
-const { useragent, super_properties, captcha } = require('../config');
-const fingerprint = require('./fingerprint');
-
-/**
- * Halt program execution for ``ms`` milliseconds.
- * @param {number} ms Milliseconds to delay. 
- */
-const delay = ms => new Promise(resolve => setTimeout(resolve, ms));
-
-const key = '6Lef5iQTAAAAAKeIvIY-DeexoO3gj7ryl9rLMEnn'; // static key
-
-/**
- * Solve a Captcha and return the key needed to bypass.
- * @param {string} verify_url The Discord verification URL
- */
-const solveCaptcha = async verify_url => {
-    const baseURL = format(`https://2captcha.com/in.php?key=%s&method=userrecaptcha&googlekey=%s&pageurl=%s&json=1`, captcha, key, verify_url);
-    const res = await fetch(baseURL);
-    const { status, request } = await res.json();
-    console.log('Sent initial request to 2captcha, received status %d.', status);
-
-    let text;
-    while(!text) {
-        console.log('No result yet, waiting 30 seconds.');
-        await delay(30000);
-
-        text = await (await fetch(format('https://2captcha.com/res.php?key=%s&id=%s&action=get', captcha, request))).text();
-        text = text === 'CAPCHA_NOT_READY' || text.substring(0, 3) !== 'OK|' ? null : text;
-    }
-
-    return text.substring(3);
-}
+import { useragent, super_properties } from '../config.js';
+import fingerprint from './fingerprint.js';
+import solveCaptcha from './util/solve_captcha.js';
 
 /**
  * Verify an email.
@@ -61,9 +31,6 @@ const verify = async (verify_url, token) => {
             'Authorization': token,
             'X-Super-Properties': super_properties,
             'X-Fingerprint': fp.fingerprint,
-            'Content-Length': Buffer.from(body).byteLength,
-            'Origin': 'https://discordapp.com',
-            'Referer': redirect_
         }
     });
 
@@ -77,4 +44,4 @@ const redirect = async url => {
     return res.url;
 }
 
-module.exports = verify;
+export default verify;