+++ src/util/ask.js
@@ -1,18 +1,23 @@
-const { createInterface } = require('readline');
-const { promisify } = require('util');
-
-const readline = createInterface({ 
-    input: process.stdin, 
-    output: process.stdout 
-});
-
-readline.question[promisify.custom] = q => new Promise(resolve => readline.question(q, resolve));
+import { createInterface } from 'readline';
+import { promisify } from 'util';
 
 /**
  * Answer a question
  * @param {string} p Question to ask.
  * @returns {Promise<string>} Answer to the question. 
  */
-const ask = p => promisify(readline.question)(p);
+const ask = async p => {
+    const readline = createInterface({ 
+        input: process.stdin, 
+        output: process.stdout 
+    });
+    
+    readline.question[promisify.custom] = q => new Promise(resolve => readline.question(q, resolve));
+    
+    const answer = await promisify(readline.question)(p);
+    readline.close();
 
-module.exports = ask;
+    return answer;
+}
+
+export default ask;